Custom keychain management is no longer supported
